requirements.txt
certifi==2020.4.5.1
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.2
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.2
shiboken2==5.14.2.1
SQLAlchemy==1.3.17
Werkzeug==1.0.1
WTForms==2.3.1
